+++ tests/sdk/merchant_card_log_test.go
@@ -1,0 +1,95 @@
+package sdk
+
+import (
+	"testing"
+	"github.com/starkbank/sdk-go/starkbank"
+	"github.com/starkbank/sdk-go/starkbank/merchantcard/log"
+	Utils "github.com/starkbank/sdk-go/tests/utils"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestMerchantCardLogGet(t *testing.T) {
+	starkbank.User = Utils.ExampleProject
+
+	limit := 1
+	var params = map[string]interface{}{}
+	params["limit"] = limit
+	
+	var logList []log.Log
+
+	logs, errorChannel := log.Query(params, nil)
+	loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			logList = append(logList, log)
+		}
+	}
+
+	retrievedLog, err := log.Get(logList[0].Id, nil)
+	if err.Errors != nil {
+		for _, erro := range err.Errors {
+			t.Errorf("code: %s, message: %s", erro.Code, erro.Message)
+		}
+	}
+	if retrievedLog.Id != logList[0].Id {
+		t.Fail()
+	}
+}
+
+func TestMerchantCardLogQuery(t *testing.T) {
+	starkbank.User = Utils.ExampleProject
+
+	var i int
+	var params = map[string]interface{}{}
+	params["limit"] = 10
+
+	logs, errorChannel := log.Query(params, nil)
+	loop:
+	for {
+		select {
+		case err := <-errorChannel:
+			if err.Errors != nil {
+				for _, e := range err.Errors {
+					t.Errorf("code: %s, message: %s", e.Code, e.Message)
+				}
+			}
+		case log, ok := <-logs:
+			if !ok {
+				break loop
+			}
+			assert.NotNil(t, log.Id)
+			i++
+		}
+	}
+	assert.Equal(t, 10, i)
+}
+func TestMerchantCardLogPage(t *testing.T) {
+	starkbank.User = Utils.ExampleProject
+
+	var ids []string
+	var params = map[string]interface{}{}
+	params["limit"] = 4
+
+	logs, cursor, err := log.Page(params, nil)
+	if err.Errors != nil {
+		for _, e := range err.Errors {
+			t.Errorf("code: %s, message: %s", e.Code, e.Message)
+		}
+	}
+	for _, log := range logs {
+		ids = append(ids, log.Id)
+		assert.NotNil(t, log.Id)
+		assert.NotNil(t, cursor)
+	}
+	assert.Len(t, ids, 4)
+}